UART driver simplified.
In order to create downlink as fast as possible without macros, a big set of functions
had been created in for each uart in 'uart.c'.
However, this set of functions made difficult to maintain the driver. Instead, just
general functions have been left and the particular uart is passed as a parameter in every execution.

Transport layers Paparazzi and Xbee have also been modified in order to follow the same procedure
(in this case there are no noticeable changes).